Fix to metric collection when VM Red
The VMMetricCollector is constructed to exit early during the metric collection
process when it detects that the VM is in a "RED" status.

The code, as previously written, will not call the
getMetricCollectorsPhaser().arriveAndDeregister() when exiting.
This in turn causes the task to never complete and write metrics to the metric
store.

To correct this, I have moved the VM status check inside the
try/catch/finally block which notifies the phaser on completion of the
task.

Ref Zendesk #269208